+++ src/Page/GroupPage/Comment/CommentInput.tsx
@@ -4,36 +4,26 @@ import { BoardsType, GetUser } from '../GroupPageBarrel';
 import { postAddComment } from '../../../API/Comment';
 import { postAddReply } from '../../../API/Comment';
 import { refreshState } from '../../../data/initialData';
+import useCommentInput from '../../../API/useCommentInput';
 
 interface commentInputProps {
-  commentInputBind: {
-    commentInput: string;
-    setCommentInput: React.Dispatch<React.SetStateAction<string>>;
-  };
-  inputToggleBind: {
-    inputToggle: boolean;
-    setInputToggle: React.Dispatch<React.SetStateAction<boolean>>;
-  };
-  isCommentBind: {
-    isComment: number;
-    setIsComment: React.Dispatch<React.SetStateAction<number>>;
-  };
-  postData: BoardsType;
   userInfo: GetUser;
+  postData: BoardsType;
   setCommentFlip: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
-const CommentInput = ({
-  commentInputBind,
-  inputToggleBind,
-  isCommentBind,
-  postData,
-  userInfo,
-  setCommentFlip,
-}: commentInputProps) => {
-  const { commentInput, setCommentInput } = commentInputBind;
-  const { inputToggle, setInputToggle } = inputToggleBind;
+const CommentInput = ({ userInfo, postData, setCommentFlip }: commentInputProps) => {
+  const {
+    isCommentBind,
+    inputToggleBind,
+    commentInputBind,
+    handleFormClickTrue,
+    handleInputChange,
+    handleFormClickFalse,
+  } = useCommentInput();
   const { isComment, setIsComment } = isCommentBind;
+  const { inputToggle, setInputToggle } = inputToggleBind;
+  const { commentInput, setCommentInput } = commentInputBind;
   const [refresh, setRefresh] = useRecoilState<number>(refreshState);
 
   const getPlaceholderText = (isComment: number, commentCount: number) => {
@@ -42,23 +32,6 @@ const CommentInput = ({
   };
 
   const placeholderText = getPlaceholderText(isComment, postData.commentCount);
-
-  // input에 들어갈 내용 CommentInput에 넣는 함수
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setCommentInput(e.target.value);
-  };
-
-  // 댓글에 붙은 input누르면 하단에 붙음
-  const handleFormClickFalse = (e: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
-    e.stopPropagation();
-    inputToggle && setInputToggle(false);
-  };
-
-  // input 바깥쪽누르면 되돌아감
-  const handleFormClickTrue = () => {
-    !inputToggle && setInputToggle(true);
-    setIsComment(0);
-  };
 
   // input 버튼 핸들러
   // 0이면 댓글추가
@@ -95,7 +68,7 @@ const CommentInput = ({
   // input에 글 적으면 화살표 노란색으로 변경
   const isTyping = commentInput.trimStart().length === 0 ? 'off' : 'on';
 
-  const CommentForm = () => {
+  const CommentForm = (): JSX.Element => {
     return (
       <CommentFormS inputToggle={inputToggle} onClick={handleFormClickFalse}>
         <InputS inputToggle={inputToggle}>
